Make attach optional in launch

+++ src/process.cpp
@@ -32,7 +32,7 @@ sdb::stop_reason::stop_reason(int wait_status) {
 
 /* fork() -> Conditional -> ptrace() -> execlp() -> wait_on_signal() */
 std::unique_ptr<sdb::process> sdb::process::launch (
-		std::filesystem::path path) {
+		std::filesystem::path path, bool debug) {
 	pipe channel(true);
 	pid_t pid = 0;
 	
@@ -43,7 +43,7 @@ std::unique_ptr<sdb::process> sdb::process::launch (
 	/* I am a child */
 	if (pid == 0) {
 		channel.close_read();
-		if (ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) {
+		if (debug and ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) {
 			exit_with_perror(channel, "Tracing failed");
 		}
 
@@ -63,8 +63,10 @@ std::unique_ptr<sdb::process> sdb::process::launch (
 	}
 
 	//Internal constructor
-	std::unique_ptr<process> proc (new process(pid, true));
-	proc->wait_on_signal();
+	std::unique_ptr<process> proc (new process(pid, true, debug));
+	if (debug) {
+		proc->wait_on_signal();
+	}
 
 	return proc;
 }
@@ -81,7 +83,7 @@ std::unique_ptr<sdb::process> sdb::process::attach(
 		error::send_errno("Could not attach");
 	}
 
-	std::unique_ptr<process> proc (new process(pid, false));
+	std::unique_ptr<process> proc (new process(pid, false, true));
 	proc->wait_on_signal();
 	return proc;
 }
@@ -107,17 +109,19 @@ sdb::stop_reason sdb::process::wait_on_signal() {
 sdb::process::~process() {
 	if (pid_ != 0) {
 		int status;
-		if (state_ == process_state::running) {
-			kill(pid_, SIGSTOP);
-			waitpid(pid_, &status, 0);
-		}
+		if(is_attached_) {
+			if (state_ == process_state::running) {
+				kill(pid_, SIGSTOP);
+				waitpid(pid_, &status, 0);
+			}
 
-		ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
-		kill(pid_, SIGCONT);
+			ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
+			kill(pid_, SIGCONT);
 
-		if (terminate_on_end_) {
-			kill(pid_, SIGKILL);
-			waitpid(pid_, &status, 0);
+			if (terminate_on_end_) {
+				kill(pid_, SIGKILL);
+				waitpid(pid_, &status, 0);
+			}
 		}
 	}
 }
